Move sort to a different place
