+++ python/taichi/tools/__init__.py
@@ -1,6 +1,5 @@
 from .image import imdisplay, imread, imresize, imshow, imwrite
 from .np2ply import PLYWriter
-from .sort import parallel_sort
 from .util import *
 # Don't import taichi_logo here which will cause circular import.
 # If you need it, just import from taichi.tools.patterns
@@ -10,5 +9,5 @@ __all__ = [
     'PLYWriter', 'VideoManager', 'imdisplay', 'imread', 'imresize', 'imshow',
     'imwrite', 'deprecated', 'warning', 'dump_dot', 'dot_to_pdf', 'obsolete',
     'get_kernel_stats', 'get_traceback', 'set_gdb_trigger',
-    'print_profile_info', 'clear_profile_info', 'parallel_sort'
+    'print_profile_info', 'clear_profile_info', 
 ]
